Adding support for declarations.
This is the easiest way for now to link externally defined libraries.
The nominal prescription is still to define global variables and functions directly.
To nudge this point, declarations are kept independent of definitions,
i.e. an identifier cannot be declared and defined (this enforces the use of
declarations for external, opaque definitions, such as the C standard library).
Moreover, multiple declarations of an identifier are not allowed.
This is meant to enforce streamlining of the calls to external entities.